refactor: 락 해제를 이벤트 리스너가 수행하도록 리팩토링
