+++ src/main/kotlin/kr/hhplus/be/server/domain/support/lock/DistributedLockClient.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.lock
+package kr.hhplus.be.server.domain.support.lock
 
 import java.util.concurrent.TimeUnit
