+++ src/main/kotlin/kr/hhplus/be/server/domain/support/lock/LockResourceManager.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.lock
+package kr.hhplus.be.server.domain.support.lock
 
 interface LockResourceManager : AutoCloseable {
     fun unlock()
