+++ src/main/kotlin/kr/hhplus/be/server/infra/lock/RedissonLockClient.kt
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.infra.lock
 
-import kr.hhplus.be.server.domain.lock.DistributedLockClient
-import kr.hhplus.be.server.domain.lock.LockResourceManager
+import kr.hhplus.be.server.domain.support.lock.DistributedLockClient
+import kr.hhplus.be.server.domain.support.lock.LockResourceManager
 import org.redisson.api.RLock
 import org.redisson.api.RedissonClient
 import java.util.concurrent.TimeUnit
